pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "config"]

[project]
name = "django-rag"
version = "0.1.0"
description = "Chat com LLM e RAG sobre base de conhecimento institucional e documentos pessoais"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web / API
    "django>=6.0,<7.0",
    # "djangorestframework>=3.15",

    # WebSocket
    # "channels>=4.1",
    # "channels-redis>=4.1",
    # "daphne>=4.1",

    # Autenticação OIDC
    "mozilla-django-oidc>=4.0",

    # Configuração / ambiente
    "django-environ>=0.11",

    # Cache / sessão
    "django-redis>=5.4",

    # Task queue
    # "celery[redis]>=5.3",

    # Banco de dados
    "psycopg[binary]>=3.1",
    # "pgvector>=0.3",

    # LangChain (chunking, integração Ollama)
    # "langchain>=0.2",
    # "langchain-community>=0.2",
    # "langchain-postgres>=0.0.9",
    # "langchain-ollama>=0.1",
    # "langchain-experimental>=0.0.60",  # SemanticChunker

    # Embeddings + reranker
    # "sentence-transformers>=3.0",

    # Extração de texto de documentos
    # "pypdf>=4.0",
    # "python-docx>=1.1",

    # Avaliação do pipeline RAG
    # "ragas>=0.1",
    # "datasets>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "pytest-asyncio",
    "factory-boy",
    "ruff",
    "mypy",
    "django-stubs",
    "djangorestframework-stubs",
    "django-debug-toolbar",
    "ipython",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-django",
    "pytest-asyncio",
    "factory-boy",
    "ruff",
    "mypy",
    "django-stubs",
    "djangorestframework-stubs",
    "django-debug-toolbar",
    "ipython",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.mypy]
python_version = "3.14"
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
strict = false

[tool.django-stubs]
django_settings_module = "config.settings.development"
